# 42_DigiConnect/back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/applications": {
      "post": {
        "operationId": "ApplicationsController_submit",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role of the caller",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreateApplicationDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApplicationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Application submitted successfully"
          }
        },
        "summary": "Submit a new application",
        "tags": [
          "applications"
        ]
      },
      "get": {
        "operationId": "ApplicationsController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "officerId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "required": false,
            "in": "header",
            "description": "ID of the caller",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "Role of the caller",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of applications"
          }
        },
        "summary": "Get all applications (paginated, filterable)",
        "tags": [
          "applications"
        ]
      }
    },
...
    "/api/v1/applications/my": {
      "get": {
        "operationId": "ApplicationsController_findMy",
        "parameters": [
          {
            "name": "x-user-id",
            "required": true,
            "in": "header",
            "description": "ID of the caller",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "Role of the caller",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of citizen applications"
          }
        },
        "summary": "Get my applications (paginated)",
        "tags": [
          "applications"
        ]
      }
    },
...
    "/api/v1/applications/officer-queue": {
      "get": {
        "operationId": "ApplicationsController_getOfficerQueue",
        "parameters": [
          {
            "name": "x-user-id",
            "required": false,
            "in": "header",
            "description": "Officer ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "Role of the caller",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Officer queue retrieved"
          }
        },
        "summary": "Get officer application queue",
        "tags": [
          "applications"
        ]
      }
    },
...
    "/api/v1/applications/{id}": {
      "get": {
        "operationId": "ApplicationsController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "Role of the caller",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Application found"
          }
        },
        "summary": "Get application by ID",
        "tags": [
          "applications"
        ]
      },
      "delete": {
        "operationId": "ApplicationsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "Role of the caller",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Application withdrawn successfully"
          }
        },
        "summary": "Withdraw an application",
        "tags": [
          "applications"
        ]
      }
    },
...
    "/api/v1/notifications": {
      "get": {
        "operationId": "NotificationsController_findMy",
        "parameters": [
          {
            "name": "x-user-id",
            "required": true,
            "in": "header",
            "description": "ID of the caller",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "Role of the caller",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of notifications"
          }
        },
        "summary": "Get all notifications for current user",
        "tags": [
          "notifications"
        ]
      },
      "post": {
        "operationId": "NotificationsController_createNotification",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role of the caller",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Notification created"
          }
        },
        "summary": "Create a manual notification",
        "tags": [
          "notifications"
        ]
      }
    },
...
    "/api/v1/notifications/read-all": {
      "patch": {
        "operationId": "NotificationsController_markAllAsRead",
        "parameters": [
          {
            "name": "x-user-id",
            "required": true,
            "in": "header",
            "description": "ID of the caller",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "Role of the caller",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "All marked as read"
          }
        },
        "summary": "Mark all notifications as read",
        "tags": [
          "notifications"
        ]
      }
    },
    "/api/v1/notifications/{id}/read": {
      "patch": {
        "operationId": "NotificationsController_markAsRead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "Role of the caller",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Notification marked as read"
          }
        },
        "summary": "Mark notification as read",
        "tags": [
          "notifications"
        ]
      }
    },
...
    "/api/v1/certificates/application/{appId}": {
      "get": {
        "operationId": "CertificatesController_getAppCertificate",
        "parameters": [
          {
            "name": "appId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get certificate metadata for an application",
        "tags": [
          "certificates"
        ]
      }
    },
    "/api/v1/grievances": {
      "post": {
        "operationId": "GrievancesController_raise",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role of the caller",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreateGrievanceDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGrievanceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Grievance raised successfully"
          }
        },
        "summary": "Raise a new grievance",
        "tags": [
          "grievances"
        ]
      },
      "get": {
        "operationId": "GrievancesController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "required": false,
            "in": "header",
            "description": "ID of the caller",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "Role of the caller",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of all grievances"
          }
        },
        "summary": "Get all grievances (paginated)",
        "tags": [
          "grievances"
        ]
      }
    },
...
    "/api/v1/grievances/{id}": {
      "get": {
        "operationId": "GrievancesController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "Role of the caller",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Grievance found"
          }
        },
        "summary": "Get grievance by ID",
        "tags": [
          "grievances"
        ]
      }
    },
...
    "/api/v1/grievances/{id}/resolve": {
      "post": {
        "operationId": "GrievancesController_resolve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "required": true,
            "in": "header",
            "description": "ID of the grievance officer",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "GRIEVANCE_OFFICER",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Explicit Grievance Resolution Action (Sections 28-31)",
        "tags": [
          "grievances"
        ]
      }
    },
    "/api/v1/services": {
      "get": {
        "operationId": "ServicesController_findActive",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of active services"
          }
        },
        "summary": "Get all active services",
        "tags": [
          "services"
        ]
      },
      "post": {
        "operationId": "ServicesController_create",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role of the caller",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateServiceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Service created successfully"
          }
        },
        "summary": "Create a new service",
        "tags": [
          "services"
        ]
      }
    },
...
    "/api/v1/services/{id}": {
      "get": {
        "operationId": "ServicesController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Service details"
          }
        },
        "summary": "Get service by ID",
        "tags": [
          "services"
        ]
      },
      "patch": {
        "operationId": "ServicesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "Role of the caller",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateServiceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Service updated successfully"
          }
        },
        "summary": "Update an existing service",
        "tags": [
          "services"
        ]
      }
    },
...
    "/api/v1/department-head/services/{id}": {
      "get": {
        "operationId": "DepartmentHeadController_getService",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get complete service specification & form configuration",
        "tags": [
          "department-head"
        ]
      }
    },
...
  },
  "info": {
    "title": "UCSDP Backend API",
    "description": "Unified Citizen Service and Delivery Platform API Documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateApplicationDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateGrievanceDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateServiceDto": {
        "type": "object",
        "properties": {}
      },
...
      "UpdateServiceDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}